Clarify README around non-Immutable object parsing
I spent a good chunk of time trying to figure out why I was getting weird errors on the strings in my state after persisting data, and finally realized it was because the transformer was trying to deal with those as well. The functionality for fixing this is already there, but it could be very helpful to future visitors to have this use case called out, as I imagine it is pretty common.